feat(sentences-per-line): add `doesEndWithIgnoredWord` function
This change exposes a new function from the `sentences-per-line` package to determine if the line of text ends with an ignored word.

It also updates both prettier-plugin-sentences-per-line to use this new function.

In order to support this, I updated the `sentences-per-line` plugin to use `publishConfig`, and referenced the package from `prettier-plugin-sentences-per-line` with workspace protocol.  That way, local development will use the source (no longer requiring a build happen), while the published package will get the properties for the built js.